frontend-infra/requirements.txt
aws-cdk-lib==2.206.0
constructs==10.4.2
boto3